% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_imgtoplt.R
\name{coord_imgtoplt}
\alias{coord_imgtoplt}
\title{coord_imgtoplt}
\usage{
coord_imgtoplt(data, scale)
}
\arguments{
\item{data}{A data frame that contains the coordinates of the points on the image}

\item{scale}{A data frame of 2 rows and 4 columns built as follow :
'x_img' is the column that contains x range on the image that correspond to the x axes of the plot
'x_plot' is the column that contains range of the x axes of the plot
'y_img' is the column that contains y range on the image that correspond to the y axes of the plot
'y_plot' is the column that contains range of the y axes of the plot}
}
\value{
The data rescaled on the original plot
}
\description{
A function to get back coordinates from the original plot according
the coordinates from the image of the plot. You'll need to know
what are the coordinates on the image of the axes.
}
