% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_line.R
\name{dist_line}
\alias{dist_line}
\title{dist_line}
\usage{
dist_line(points, lines)
}
\arguments{
\item{points}{A data frame that contains the coordinates of the points}

\item{lines}{A data frame that contains the coordinates of the points that defines the path}
}
\value{
A list of 3 elements. 'dist' is the minimum distance between each points and the lines,
        'proj' is the projection from the point on the line,
        'which_line' is the index of the line where the point is projected.
}
\description{
Calculate distance between a line and set of points; several lines = path
}
