% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawYpath.R
\docType{package}
\name{drawYpath}
\alias{drawYpath}
\title{drawYpath}
\description{
This package contains a shiny app that allow the user to draw path on a plot thanks to plotly.
Then, it finds the points next to this path and plot a heatmap of those points.
This app can be useful when the user want to find its own trajectories, to keep a more biological view
than just  use an algorithm.
}
\author{
Marc-Antoine GERAULT \email{marco.gerault@gmail.com}
}
