% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_plot.R
\name{int_plot}
\alias{int_plot}
\title{int_plot}
\usage{
int_plot(data, img, tit = "t-SNE", src = "M", max_points = 3000)
}
\arguments{
\item{data}{A data frame that contains the data. For now, needs to have 'tSNE1', 'tSNE2', 'pred'
and 'index' columns. The 'tSNE' axes are the y and x axes respectively, 'pred' is the color
and 'index' is the row names.}

\item{img}{The image of the plot. Need to be an 'Image' object from \code{EBImage} package}

\item{tit}{The title of the plot}

\item{src}{The source of the plotly plot, default is "M"}

\item{max_points}{The maximum number of points you want to print before ploting the image of the plot}
}
\value{
An interactive plot from plotly on which can draw path
}
\description{
Plot an interactive graph of the direct data if the number of points to print is
not to high, else it plot the image of the plot.
}
