% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_path.R
\name{spline_path}
\alias{spline_path}
\title{spline_path}
\usage{
spline_path(x, y, n, t = seq_along(x), method = "natural")
}
\arguments{
\item{x}{A vector that contains the x coordinates from the points that defines the path}

\item{y}{A vector that contains the y coordinates from the points that defines the path}

\item{n}{The number of points you want to keep after spline}

\item{t}{The order of the points that defines the path; default is \code{seq_along(x)}}

\item{method}{Method from spline function. It precise the type of spline to use.
Possible values are 'fmm', 'natural', 'periodic', 'monoH.FC' and 'hyman'.}
}
\value{
A data frame of the new points that defines the path
}
\description{
A function that simplify a given path; remove points by splining the curve.
}
